run Keycloak during the integration tests phase
