+++ vertx-auth-oauth2/src/test/java/io/vertx/ext/auth/test/oauth2/OAuth2KeycloakIT.java
@@ -6,13 +6,9 @@ import io.vertx.ext.auth.oauth2.providers.KeycloakAuth;
 import io.vertx.test.core.VertxTestBase;
 import org.junit.Test;
 
-import java.net.HttpURLConnection;
-import java.net.URL;
-
-public class OAuth2KeycloakTest extends VertxTestBase {
+public class OAuth2KeycloakIT extends VertxTestBase {
 
   private OAuth2Auth oauth2;
-  private boolean isKeycloakAvailable;
 
   // Set the client credentials and the OAuth2 server
   final JsonObject credentials = new JsonObject(
@@ -30,26 +26,11 @@ public class OAuth2KeycloakTest extends VertxTestBase {
   public void setUp() throws Exception {
     super.setUp();
     oauth2 = OAuth2Auth.createKeycloak(vertx, OAuth2FlowType.PASSWORD, credentials);
-
-    try {
-      URL url = new URL("http://localhost:8888/auth");
-      HttpURLConnection httpConn = (HttpURLConnection) url.openConnection();
-      if (200 == httpConn.getResponseCode()) {
-        isKeycloakAvailable = true;
-      }
-    } catch (Throwable e) {
-      System.err.println("!!! Keycloak is not available");
-      // assume tests pass as keycloak is not available
-      isKeycloakAvailable = false;
-    }
   }
 
   @Test
   public void testFullCycle() {
-    if (!isKeycloakAvailable) {
-      testComplete();
-      return;
-    }
+
     oauth2.authenticate(new JsonObject().put("username", "user").put("password", "password"), res -> {
       if (res.failed()) {
         fail(res.cause().getMessage());
@@ -88,10 +69,7 @@ public class OAuth2KeycloakTest extends VertxTestBase {
 
   @Test
   public void testLogout() {
-    if (!isKeycloakAvailable) {
-      testComplete();
-      return;
-    }
+
     oauth2.authenticate(new JsonObject().put("username", "user").put("password", "password"), res -> {
       if (res.failed()) {
         fail(res.cause().getMessage());
@@ -121,11 +99,7 @@ public class OAuth2KeycloakTest extends VertxTestBase {
 
   @Test
   public void testDecodeShouldFail() throws Exception {
-    super.setUp();
-    if (!isKeycloakAvailable) {
-      testComplete();
-      return;
-    }
+
     oauth2 = KeycloakAuth.create(vertx, OAuth2FlowType.AUTH_CODE, credentials);
     oauth2.decodeToken("borked", res1 -> {
       if (res1.failed()) {
@@ -140,11 +114,7 @@ public class OAuth2KeycloakTest extends VertxTestBase {
 
   @Test
   public void testDecodeShouldPass() throws Exception {
-    super.setUp();
-    if (!isKeycloakAvailable) {
-      testComplete();
-      return;
-    }
+
     oauth2 = KeycloakAuth.create(vertx, OAuth2FlowType.PASSWORD, credentials);
 
     oauth2.loadJWK(v -> {
@@ -173,4 +143,26 @@ public class OAuth2KeycloakTest extends VertxTestBase {
 
     await();
   }
+
+  @Test
+  public void testLoadJWK2() {
+    JsonObject config = new JsonObject("{\n" +
+      "  \"realm\": \"master\",\n" +
+      "  \"auth-server-url\": \"http://localhost:8888/auth\",\n" +
+      "  \"ssl-required\": \"external\",\n" +
+      "  \"resource\": \"test\",\n" +
+      "  \"credentials\": {\n" +
+      "    \"secret\": \"b0568625-a482-45d8-af8b-27beba502ed3\"\n" +
+      "  }\n" +
+      "}");
+
+    OAuth2Auth oauth2 = KeycloakAuth.create(vertx, config);
+
+    oauth2.loadJWK(load -> {
+      assertFalse(load.failed());
+      testComplete();
+    });
+    await();
+  }
+
 }
